split the parsing flow into multiple files
